UMDCTF2023/forensics/yara-trainer-gym/dev/yara-mart-site/yara_mart_site/requirements.txt
click==8.1.3 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
flask==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
yara-python==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
